refactoring controllers and dtos
