--- src/main/java/com/ottistech/indespensa/api/ms_indespensa/dto/LoginUserDTO.java
@@ -1,11 +1,0 @@
-package com.ottistech.indespensa.api.ms_indespensa.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotNull;
-
-public record LoginUserDTO(
-        @Email(message = "Email isn't right") @NotNull(message = "Field email is required") String email,
-        @NotNull(message = "Field password is required") String password
-) {
-
-}
