--- src/main/java/com/ottistech/indespensa/api/ms_indespensa/dto/UpdatePantryItemDTO.java
@@ -1,8 +1,0 @@
-package com.ottistech.indespensa.api.ms_indespensa.dto;
-
-public record UpdatePantryItemDTO (
-    Long pantryItemId,
-    Integer pantryAmount
-) {
-
-}
