+++ src/main/java/com/ottistech/indespensa/api/ms_indespensa/dto/request/AddShopItemDTO.java
@@ -8,8 +8,8 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
 public record AddShopItemDTO(
-        Long productId,
-        @NotNull(message = "Field amount is required") @Min(value = 1, message = "Amount must be at least 1") Integer amount
+        @NotNull(message = "Field productId is required") @Min(value = 1, message = "Field productId must be at least 1") Long productId,
+        @NotNull(message = "Field amount is required") @Min(value = 1, message = "Field amount must be at least 1") Integer amount
 ) {
 
     public ShopItem toShopItem(User user, Product product, AddShopItemDTO shopItemDTO) {
