+++ src/main/java/com/ottistech/indespensa/api/ms_indespensa/dto/request/PantryItemCreateDTO.java
@@ -1,5 +1,6 @@
-package com.ottistech.indespensa.api.ms_indespensa.dto;
+package com.ottistech.indespensa.api.ms_indespensa.dto.request;
 
+import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
 import java.math.BigDecimal;
@@ -10,12 +11,12 @@ public record PantryItemCreateDTO(
         @NotNull(message = "Field productName is required") String productName,
         @NotNull(message = "Field productDescription is required") String productDescription,
         String productImageUrl,
-        @NotNull(message = "Field productAmount is required") BigDecimal productAmount,
+        @NotNull(message = "Field productAmount is required") @Min(value = 0, message = "Field productAmount should be at least a positive number") BigDecimal productAmount,
         @NotNull(message = "Field productUnit is required") String productUnit,
         @NotNull(message = "Field foodName is required") String foodName,
         @NotNull(message = "Field categoryName is required") String categoryName,
         @NotNull(message = "Field brandName is required") String brandName,
-        @NotNull(message = "Field pantryAmount is required") Integer pantryAmount,
+        @NotNull(message = "Field pantryAmount is required") @Min(value = 1, message = "Field productAmount should be at least 1") Integer pantryAmount,
         @NotNull(message = "Field validityDate is required") LocalDate validityDate
 ) {
 }
