+++ src/main/java/com/ottistech/indespensa/api/ms_indespensa/dto/response/PantryItemPartialDTO.java
@@ -1,4 +1,4 @@
-package com.ottistech.indespensa.api.ms_indespensa.dto;
+package com.ottistech.indespensa.api.ms_indespensa.dto.response;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
