+++ src/main/java/com/ottistech/indespensa/api/ms_indespensa/dto/response/UserCredentialsResponseDTO.java
@@ -1,4 +1,4 @@
-package com.ottistech.indespensa.api.ms_indespensa.dto;
+package com.ottistech.indespensa.api.ms_indespensa.dto.response;
 
 public record UserCredentialsResponseDTO(
     Long userId,
