+++ src/main/java/com/ottistech/indespensa/api/ms_indespensa/dto/response/UserFullInfoResponseDTO.java
@@ -1,4 +1,4 @@
-package com.ottistech.indespensa.api.ms_indespensa.dto;
+package com.ottistech.indespensa.api.ms_indespensa.dto.response;
 
 import java.util.Date;
