+++ src/main/java/com/ottistech/indespensa/api/ms_indespensa/service/CategoryService.java
@@ -1,10 +1,11 @@
 package com.ottistech.indespensa.api.ms_indespensa.service;
 
-import com.ottistech.indespensa.api.ms_indespensa.exception.CategoryNotFoundException;
 import com.ottistech.indespensa.api.ms_indespensa.model.Category;
 import com.ottistech.indespensa.api.ms_indespensa.repository.CategoryRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -24,8 +25,9 @@ public class CategoryService {
         }
 
         if(result.isEmpty()) {
-            throw new CategoryNotFoundException("No categories matching the pattern");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No categories matching the pattern");
         }
+
         return result.stream()
                 .map(Category::getCategoryName)
                 .collect(Collectors.toList());
